Move browser methods into PlatformUI CommonActions context
