--- Features/Context/Override/CommonActions.php
@@ -1,79 +1,0 @@
-<?php
-
-/**
- * File containing Overriden Functions for context PlatformUI.
- *
- * @copyright Copyright (C) eZ Systems AS. All rights reserved.
- * @license For full copyright and license information view LICENSE file distributed with this source code.
- * @version //autogentag//
- */
-namespace EzSystems\PlatformUIBundle\Features\Context\Override;
-
-trait CommonActions
-{
-    /**
-     * @Overrride
-     * @When I fill in :field with :value
-     * @When I set :field as empty
-     *
-     * Overriden fill fields function
-     * Spin function make it possible to retry in case of failure
-     */
-    public function fillFieldWithValue($field, $value = '')
-    {
-        $fieldNode = $this->spin(
-            function () use ($field) {
-                $fieldNode = $this->getSession()->getPage()->findField($field);
-                if ($fieldNode == null) {
-                    throw new \Exception('Field not found');
-                }
-
-                return $fieldNode;
-            }
-        );
-
-        $this->spin(
-            function () use ($fieldNode, $field, $value) {
-                // make sure any autofocus elements don't mis-behave when setting value
-                $fieldNode->blur();
-                usleep(10 * 1000);
-                $fieldNode->focus();
-                usleep(10 * 1000);
-
-                // setting value on pre-filled inputs can cause issues, clearing before
-                $fieldNode->setValue('');
-                $fieldNode->setValue($value);
-
-                // verication that the field was really filled in correctly
-                $this->sleep();
-                $check = $this->getSession()->getPage()->findField($field)->getValue();
-                if ($check != $value) {
-                    throw new \Exception('Failed to set the field value: ' . $check);
-                }
-
-                return true;
-            }
-        );
-    }
-
-    /**
-     * @Override
-     * @Then I (should) see :title title/topic
-     */
-    public function iSeeTitle($title)
-    {
-        $page = $this->getSession()->getPage();
-        $this->spin(
-            function () use ($title, $page) {
-                $titleElements = $page->findAll('css', 'h1, h2, h3');
-                foreach ($titleElements as $titleElement) {
-                    $elementText = $titleElement->getText();
-                    if ($elementText == $title) {
-                        return $titleElement;
-                    }
-                }
-                throw new \Exception("Title '$title' not found");
-            }
-        );
-    }
-}
